+++ .env.development
@@ -1,0 +1,1 @@
+REACT_APP_API = http://localhost:7123
