+++ src/components/Home/cards/new-item/NewItemCards.scoped.scss
