+++ src/components/Home/carousels/category/carousel.scoped.scss
