+++ src/components/Home/carousels/trending/carousel.scoped.scss
