+++ src/components/Logins/Form.scoped.scss
@@ -1,0 +1,44 @@
+form {
+	margin-top: 40px;
+	input {
+		border: 1px solid #9b9b9b;
+		box-sizing: border-box;
+		filter: drop-shadow(0px 1px 8px rgba(0, 0, 0, 0.05));
+		border-radius: 4px;
+	}
+
+	.btn-box {
+		.btn-login {
+			height: 48px;
+			width: 390px;
+			background: #db3022;
+			border-radius: 25px;
+			border: none;
+		}
+	}
+	p {
+		font-style: normal;
+		font-weight: normal;
+		font-size: 14px;
+		line-height: 14px;
+		color: #db3022;
+		display: flex;
+		left: 280px;
+		margin-top: 30px;
+		margin-bottom: 30px;
+		position: relative;
+	}
+	.error {
+		font-size: 12px;
+		margin: 0px;
+	}
+}
+h2 {
+	font-style: normal;
+	font-weight: normal;
+	font-size: 14px;
+	line-height: 14px;
+	margin-top: 40px;
+	text-decoration: none;
+	color: black;
+}
