+++ src/components/Update-product/UpdateProduct.scoped.scss
@@ -38,6 +38,9 @@
 		}
 		background-color: #e5e5e5;
 	}
+	.id {
+		display: none;
+	}
 }
 
 .photo-of-goods {
