--- src/components/image-upload/styles.scss
@@ -1,17 +1,0 @@
-.image-item {
-	display: flex;
-	flex-direction: row;
-
-	margin: 10px 0;
-}
-.image-item__btn-wrapper {
-	display: flex;
-	flex-direction: column;
-	margin-left: 10px;
-}
-button {
-	height: 40px;
-}
-.goods {
-	width: 700px;
-}
