--- src/components/logins/login.scoped.css
@@ -1,146 +1,0 @@
-body {
-	font-family: metropolis;
-	text-align: center;
-	margin: 0;
-	font-size: 12;
-}
-
-.form-box {
-	width: 350px;
-	height: 526px;
-	position: relative;
-	margin: 6% auto;
-	background: transparent;
-	padding: 5px;
-	overflow: hidden;
-}
-
-.button-box {
-	top: 40px;
-	width: 220px;
-	margin: 35px auto;
-	position: relative;
-	background-color: white;
-	border-radius: 4px;
-	border: 1px solid #9b9b9b;
-}
-
-.toggle-button {
-	padding: 10px 30px;
-	cursor: pointer;
-	background: transparent;
-	border: 0;
-	outline: none;
-	position: relative;
-	color: #9b9b9b;
-}
-
-#button {
-	top: 0;
-	left: 0;
-	position: absolute;
-	width: 110px;
-	height: 100%;
-	background: #db3022;
-	border-radius: 4px;
-	-webkit-transition: 0.5s;
-	transition: 0.5s;
-}
-
-.input-group {
-	top: 225px;
-	position: absolute;
-	width: 280px;
-	-webkit-transition: 0.5s;
-	transition: 0.5s;
-}
-
-.input-field {
-	width: 300px;
-	padding: 17px 0;
-	margin: 5px;
-	border: 1px solid #9b9b9b;
-	right: 40px;
-	left: 60;
-	background: transparent;
-	border-radius: 4px;
-	position: relative;
-	padding-left: 20px;
-}
-
-.submit-button {
-	top: 50px;
-	right: 30px;
-	padding: 10px 142px;
-	cursor: pointer;
-	display: block;
-	margin: auto;
-	background: #db3022;
-	border: 0;
-	border-radius: 30px;
-	outline: none;
-	position: relative;
-	color: white;
-}
-
-#login-customer {
-	left: 50px;
-}
-
-#login-seller {
-	left: 450px;
-}
-
-#tog-cust {
-	color: white;
-	right: 5px;
-}
-
-#tog-seller {
-	color: #9b9b9b;
-	right: 5px;
-}
-
-.logo {
-	width: 135px;
-	height: 50px;
-	position: relative;
-	display: block;
-	margin-left: auto;
-	margin-right: auto;
-	width: 50%;
-}
-
-.title {
-	top: 30px;
-	position: relative;
-	font-size: 10px;
-	font-weight: bold;
-}
-
-.text-forgot {
-	position: relative;
-	top: 165px;
-	text-align: right;
-	font-size: 12px;
-	color: red;
-	right: 20px;
-	text-decoration: none;
-}
-
-.tokopedia {
-	position: relative;
-	top: 250px;
-	font-size: 12px;
-}
-
-.register {
-	color: red;
-	text-decoration: none;
-}
-
-.link-forgot-pass {
-	color: red;
-	text-decoration: none;
-}
-/*# sourceMappingURL=style-login.css.map */
