+++ src/components/navbar/NavbarAfterLogin.jsx
@@ -33,51 +33,69 @@ const NavbarAfterLogin = () => {
 					<Navbar.Toggle aria-controls="navbar-collapse" />
 
 					<Container className="navbar-inside">
-						<Navbar.Collapse id="basic-navbar-nav">
-							<Container className="tools">
-								<Container className="search-bar">
-									<Form className="d-flex search-box bg-transparent ">
-										<FormControl
-											type="search"
-											placeholder="Search"
-											className="mr-2"
-											aria-label="Search"
-										/>
-									</Form>
-									<Nav.Link href="#home">
-										<FaSearch className="search-icon" />
-									</Nav.Link>
+						<Navbar.Collapse id="navbar-collapse-before-login">
+							<div className="navbar-mobile">
+								<ul>
+									<li>
+										<Nav.Link href="#home">SignUp</Nav.Link>
+									</li>
+									<li>
+										<Nav.Link href="#home">Login</Nav.Link>
+									</li>
+									<li>
+										<Nav.Link href="#home">Cart</Nav.Link>
+									</li>
+								</ul>
+							</div>
+						</Navbar.Collapse>
+
+						<Navbar.Collapse className="basic-navbar" id="basic-navbar-nav">
+							<Container className="navbar-inside">
+								<Container className="tools">
+									<Container className="search-bar">
+										<Form className="d-flex search-box bg-transparent ">
+											<FormControl
+												type="search"
+												placeholder="Search"
+												className="mr-2"
+												aria-label="Search"
+											/>
+										</Form>
+										<Nav.Link href="#home">
+											<FaSearch className="search-icon" />
+										</Nav.Link>
+									</Container>
+									<Container className="filter-box">
+										<Nav.Link href="#home">
+											<FaFilter className="filter-icon" />
+										</Nav.Link>
+									</Container>
 								</Container>
-								<Container className="filter-box">
-									<Nav.Link href="#home">
-										<FaFilter className="filter-icon" />
-									</Nav.Link>
+								<Container>
+									<Nav className="me-auto menu-after-login">
+										<Nav.Link>
+											<Container className="menu-icon">
+												<FaShoppingCart />
+											</Container>
+										</Nav.Link>
+										<Nav.Link>
+											<Container className="menu-icon">
+												<FaBell />
+											</Container>
+										</Nav.Link>
+										<Nav.Link>
+											<Container className="menu-icon">
+												<FaEnvelope />
+											</Container>
+										</Nav.Link>
+										<Nav.Link className="photo-box">
+											<img
+												src="https://res.cloudinary.com/calvin-cloud/image/upload/v1627339637/Front%20End/profile_yzozml.jpg"
+												alt="profile"
+											/>
+										</Nav.Link>
+									</Nav>
 								</Container>
-							</Container>
-							<Container>
-								<Nav className="me-auto menu-after-login">
-									<Nav.Link>
-										<Container className="menu-icon">
-											<FaShoppingCart />
-										</Container>
-									</Nav.Link>
-									<Nav.Link>
-										<Container className="menu-icon">
-											<FaBell />
-										</Container>
-									</Nav.Link>
-									<Nav.Link>
-										<Container className="menu-icon">
-											<FaEnvelope />
-										</Container>
-									</Nav.Link>
-									<Nav.Link className="photo-box">
-										<img
-											src="https://res.cloudinary.com/calvin-cloud/image/upload/v1627339637/Front%20End/profile_yzozml.jpg"
-											alt="profile"
-										/>
-									</Nav.Link>
-								</Nav>
 							</Container>
 						</Navbar.Collapse>
 					</Container>
