--- src/http-common.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-export default axios.create({
-	baseURL: 'http://localhost:8080/api',
-	headers: {
-		'Content-type': 'application/json',
-	},
-});
