+++ src/pages/AddProduct/AddProduct.jsx
@@ -1,0 +1,13 @@
+import NavbarAfterLogin from '../../components/navbar/NavbarAfterLogin';
+import SellingProduct from '../../components/selling-product/SellingProduct';
+
+const AddProduct = () => {
+	return (
+		<div>
+			<NavbarAfterLogin />
+			<SellingProduct />
+		</div>
+	);
+};
+
+export default AddProduct;
