+++ src/pages/EditProduct/EditProduct.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import NavbarAfterLogin from '../../components/navbar/NavbarAfterLogin';
+import UpdateProduct from '../../components/Update-product/UpdateProduct';
+
+const EditProduct = () => {
+	return (
+		<div>
+			<NavbarAfterLogin />
+			<UpdateProduct />
+		</div>
+	);
+};
+
+export default EditProduct;
