+++ src/pages/Home/Home.scoped.scss
@@ -1,0 +1,3 @@
+.wrapper {
+	background-color: #f9f9f9;
+}
