+++ src/pages/ProductInventory/ProductInventory.scope.scss
@@ -1,0 +1,14 @@
+.big-container {
+	display: flex;
+
+	.left-container {
+		width: 30%;
+		margin-top: 50px;
+	}
+	.right-container {
+		padding-top: 50px;
+		margin-top: 50px;
+		background: #f5f5f5;
+		padding-bottom: 114px;
+	}
+}
