+++ src/pages/ProductPreview/ProductPreview.jsx
@@ -1,0 +1,15 @@
+import ProductPage from '../../components/Product/Product';
+
+import React from 'react';
+import NavbarInit from '../../components/navbar/Navbar';
+
+const ProductPreview = () => {
+	return (
+		<div>
+			<NavbarInit />
+			<ProductPage />
+		</div>
+	);
+};
+
+export default ProductPreview;
