+++ src/pages/index.js
@@ -1,0 +1,8 @@
+import Login from './Login/Login';
+import Home from './Home/Home';
+import BagPage from './Bag/BagPage';
+import AddProduct from './AddProduct/AddProduct';
+import ProductPreview from './ProductPreview/ProductPreview';
+import EditProduct from './EditProduct/EditProduct';
+
+export { Login, Home, BagPage, AddProduct, ProductPreview, EditProduct };
